Improves doc and changes any references of `proxy` to `adapter`.
Improves the documentation for better understanding of how the API works and made some refactoring
for improve the modularity and build performance.

The idea of `proxy` was changed for `adapter` because the adapter gives the idea of exposing a
different interface from an existent API (IPlugin), instead of keeping the same interface of an
existent API, which the proxy does. So it uses the interface of `IPlugin`, but improving with
new functionality.